+++ index.html
@@ -48,7 +48,7 @@
                         rows="4"
                         cols="7"
                         resize="false"
-                        placeholder="#C4B9B3  #434343  #657189"
+                        placeholder="#C4B9B3 #434343 #657189"
                         pattern="^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})(\n ?#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3}))*$"
                         title="Enter HEX colors separated by commas"
                     ></textarea>
