--- src/config.ts
@@ -1,4 +1,0 @@
-export const GRADIENT_SIZE = '2';
-export const CONTRAST_THRESHOLD = '2.5';
-export const SHUFFLE_PLATES = false;
-export const PALETTE = '#C4B9B3\n#434343\n#657189';
