+++ src/utils/contrast.ts
@@ -1,27 +1,16 @@
-export function isLowContrastGradient(
+export function isLowContrast(
     hex1: string,
     hex2: string[],
-    CONTRAST_THRESHOLD: number
+    threshold: number
 ) {
     const rgb1 = hexToRgb(hex1);
 
     for (let i = 0; i < hex2.length; i++) {
-        if (contrast(rgb1, hexToRgb(hex2[i])) < CONTRAST_THRESHOLD) {
+        if (contrast(rgb1, hexToRgb(hex2[i])) < threshold) {
             return true;
         }
     }
     return false;
-}
-
-export function isLowContrast(
-    hex1: string,
-    hex2: string,
-    CONTRAST_THRESHOLD: number
-) {
-    const rgb1 = hexToRgb(hex1);
-    const rgb2 = hexToRgb(hex2);
-
-    return contrast(rgb1, rgb2) < CONTRAST_THRESHOLD;
 }
 
 function hexToRgb(hex: string): number[] {
