+++ src/utils/utils.ts
@@ -1,0 +1,8 @@
+export function shuffle<T>(array: T[]): T[] {
+    // Fisher-Yates (Durstenfeld) shuffle algorithm
+    for (let i = array.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [array[i], array[j]] = [array[j], array[i]]; // Swap elements
+    }
+    return array;
+}
